// frontend/src/api/models/tipoDispositivo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * backend-node
 * OpenAPI spec version: 1.0.0
 */

export type TipoDispositivo = typeof TipoDispositivo[keyof typeof TipoDispositivo];


export const TipoDispositivo = {
  TELEFONO: 'TELEFONO',
  TABLET: 'TABLET',
  COMPUTER: 'COMPUTER',
} as const;